Cleanup and extend unit tests
