+++ src/core/test/Eryph.VmManagement.Test/CatletFeedingTests.cs
@@ -25,14 +25,13 @@ namespace Eryph.VmManagement.Test;
 public class CatletFeedingTests
 {
     private readonly Mock<ILocalGenepoolReader> _genepoolReaderMock = new();
-    // TODO fix me
-    private readonly HashMap<GeneIdentifier, Architecture> _resolvedGenes = new();
 
-    [Theory]
-    [InlineData(null)]
-    [InlineData("gene:acme/acme-os/1.0:catlet")]
+    [Theory, CombinatorialData]
     public void Feed_AllFoodShouldBeEaten_EatsAllFood(
-        string? catletSource)
+        [CombinatorialValues(null, "gene:acme/acme-os/1.0:catlet")]
+        string? catletSource,
+        [CombinatorialValues("any", "hyperv/any", "hyperv/amd64")]
+        string architecture)
     {
         var config = new CatletConfig
         {
@@ -52,9 +51,9 @@ public class CatletFeedingTests
             ],
         };
 
-        ArrangeFood();
+        var resolvedGenes = ArrangeFood(architecture);
 
-        var result = CatletFeeding.Feed(config, Empty, _genepoolReaderMock.Object);
+        var result = CatletFeeding.Feed(config, resolvedGenes, _genepoolReaderMock.Object);
 
         result.Should().BeRight().Which.Fodder.Should().SatisfyRespectively(
             fodder =>
@@ -77,11 +76,12 @@ public class CatletFeedingTests
             });
     }
 
-    [Theory]
-    [InlineData(null)]
-    [InlineData("gene:acme/acme-os/1.0:catlet")]
+    [Theory, CombinatorialData]
     public void Feed_SomeFoodShouldBeEaten_EatsOnlySpecificFood(
-        string? catletSource)
+        [CombinatorialValues(null, "gene:acme/acme-os/1.0:catlet")]
+        string? catletSource,
+        [CombinatorialValues("any", "hyperv/any", "hyperv/amd64")]
+        string architecture)
     {
         var config = new CatletConfig
         {
@@ -102,9 +102,9 @@ public class CatletFeedingTests
             ],
         };
 
-        ArrangeFood();
+        var resolvedGenes = ArrangeFood(architecture);
 
-        var result = CatletFeeding.Feed(config, Empty, _genepoolReaderMock.Object);
+        var result = CatletFeeding.Feed(config, resolvedGenes, _genepoolReaderMock.Object);
 
         result.Should().BeRight().Which.Fodder.Should().SatisfyRespectively(
             fodder =>
@@ -126,7 +126,9 @@ public class CatletFeedingTests
         [CombinatorialValues(null, "food1")]
         string? foodName,
         [CombinatorialValues(null, "gene:acme/acme-os/1.0:catlet")]
-        string? catletSource)
+        string? catletSource,
+        [CombinatorialValues("any", "hyperv/any", "hyperv/amd64")]
+        string architecture)
     {
         var config = new CatletConfig
         {
@@ -152,9 +154,9 @@ public class CatletFeedingTests
             ],
         };
 
-        ArrangeFood();
+        var resolvedGenes = ArrangeFood(architecture);
 
-        var result = CatletFeeding.Feed(config, Empty, _genepoolReaderMock.Object);
+        var result = CatletFeeding.Feed(config, resolvedGenes, _genepoolReaderMock.Object);
 
         result.Should().BeRight().Which.Fodder.Should().SatisfyRespectively(
             fodder =>
@@ -170,7 +172,9 @@ public class CatletFeedingTests
         [CombinatorialValues(null, "food2")]
         string? foodName,
         [CombinatorialValues(null, "gene:acme/acme-os/1.0:catlet")]
-        string? catletSource)
+        string? catletSource,
+        [CombinatorialValues("any", "hyperv/any", "hyperv/amd64")]
+        string architecture)
     {
         var config = new CatletConfig
         {
@@ -197,9 +201,9 @@ public class CatletFeedingTests
             ],
         };
 
-        ArrangeFood();
+        var resolvedGenes = ArrangeFood(architecture);
 
-        var result = CatletFeeding.Feed(config, Empty, _genepoolReaderMock.Object);
+        var result = CatletFeeding.Feed(config, resolvedGenes, _genepoolReaderMock.Object);
 
         result.Should().BeRight().Which.Fodder.Should().SatisfyRespectively(
             fodder =>
@@ -216,8 +220,10 @@ public class CatletFeedingTests
             });
     }
 
-    [Fact]
-    public void Feed_FoodExistsMultipleTimes_FoodIsMerged()
+    [Theory, CombinatorialData]
+    public void Feed_FoodExistsMultipleTimes_FoodIsMerged(
+        [CombinatorialValues("any", "hyperv/any", "hyperv/amd64")]
+        string architecture)
     {
         var config = new CatletConfig
         {
@@ -252,9 +258,9 @@ public class CatletFeedingTests
             ],
         };
 
-        ArrangeFood();
+        var resolvedGenes = ArrangeFood(architecture);
 
-        var result = CatletFeeding.Feed(config, Empty, _genepoolReaderMock.Object);
+        var result = CatletFeeding.Feed(config, resolvedGenes, _genepoolReaderMock.Object);
         result.Should().BeRight().Which.Fodder.Should().SatisfyRespectively(
             fodder =>
             {
@@ -282,12 +288,13 @@ public class CatletFeedingTests
             });
     }
 
-    private void ArrangeFood()
+    private Seq<UniqueGeneIdentifier> ArrangeFood(string architecture)
     {
-        _genepoolReaderMock.SetupGenesetReferences([]);
+        _genepoolReaderMock.SetupGenesetReferences();
 
         _genepoolReaderMock.SetupFodderGene(
             "gene:acme/acme-tools/1.0:test-fodder",
+            architecture,
             new FodderGeneConfig()
             {
                 Name = "test-fodder",
@@ -312,12 +319,20 @@ public class CatletFeedingTests
                     }
                 ]
             });
+
+        return Seq1(
+            new UniqueGeneIdentifier(
+                GeneType.Fodder,
+                GeneIdentifier.New("gene:acme/acme-tools/1.0:test-fodder"),
+                Architecture.New(architecture)));
     }
 
     [Theory, CombinatorialData]
     public void Feed_FodderVariableIsBound_BoundVariableIsUsed(
         bool? catletSecret,
-        bool? geneSecret)
+        bool? geneSecret,
+        [CombinatorialValues("any", "hyperv/any", "hyperv/amd64")]
+        string architecture)
     {
         var config = new CatletConfig
         {
@@ -343,6 +358,7 @@ public class CatletFeedingTests
         _genepoolReaderMock.SetupGenesetReferences();
         _genepoolReaderMock.SetupFodderGene(
             "gene:acme/acme-tools/1.0:test-fodder",
+            architecture,
             new FodderGeneConfig
             {
                 Name = "test-fodder",
@@ -367,7 +383,13 @@ public class CatletFeedingTests
                 ],
             });
 
-        var result = CatletFeeding.Feed(config, Empty, _genepoolReaderMock.Object);
+        var resolvedGenes = Seq1(
+            new UniqueGeneIdentifier(
+                GeneType.Fodder,
+                GeneIdentifier.New("gene:acme/acme-tools/1.0:test-fodder"),
+                Architecture.New(architecture)));
+
+        var result = CatletFeeding.Feed(config, resolvedGenes, _genepoolReaderMock.Object);
 
         var newConfig = result.Should().BeRight().Subject;
 
@@ -454,7 +476,33 @@ public class CatletFeedingTests
     }
 
     [Fact]
-    public void Feed_FodderWithUnresolvedSource_ReturnsError()
+    public void Feed_FodderWithUnresolvedGeneInSource_ReturnsError()
+    {
+        var config = new CatletConfig
+        {
+            Name = "test",
+            Fodder =
+            [
+                new FodderConfig()
+                {
+                    Source = "gene:acme/acme-tools/1.0:test-fodder",
+                },
+            ],
+        };
+
+        _genepoolReaderMock.SetupGenesetReferences(
+            ("acme/acme-tools/latest", "acme/acme-tools/1.0"));
+
+        var result = CatletFeeding.Feed(config, Empty, _genepoolReaderMock.Object);
+
+        var error = result.Should().BeLeft().Subject;
+        error.Message.Should().Be("Could not expand the fodder gene 'gene:acme/acme-tools/1.0:test-fodder'.");
+        error.Inner.Should().BeSome().Which.Message
+            .Should().Be("The gene 'gene:acme/acme-tools/1.0:test-fodder' has not been correctly resolved. This should not happen.");
+    }
+
+    [Fact]
+    public void Feed_FodderWithUnresolvedGeneSetInSource_ReturnsError()
     {
         var config = new CatletConfig
         {
@@ -479,8 +527,10 @@ public class CatletFeedingTests
             .Should().Be("The gene 'gene:acme/acme-tools/latest:test-fodder' is an unresolved reference. This should not happen.");
     }
 
-    [Fact]
-    public void Feed_FodderNotInGenepool_ReturnsError()
+    [Theory, CombinatorialData]
+    public void Feed_FodderNotInGenepool_ReturnsError(
+        [CombinatorialValues("any", "hyperv/any", "hyperv/amd64")]
+        string architecture)
     {
         var config = new CatletConfig
         {
@@ -496,14 +546,18 @@ public class CatletFeedingTests
 
         _genepoolReaderMock.SetupGenesetReferences();
         _genepoolReaderMock.Setup(m => m.ReadGeneContent(
-                // TODO fix me
                 new UniqueGeneIdentifier(
                     GeneType.Fodder,
                     GeneIdentifier.New("gene:acme/acme-tools/1.0:test-fodder"),
-                    Architecture.New("any"))))
+                    Architecture.New(architecture))))
             .Returns(Error.New("Gene 'gene:acme/acme-tools/1.0:test-fodder' does not exist in local genepool."));
 
-        var result = CatletFeeding.Feed(config, Empty, _genepoolReaderMock.Object);
+        var resolvedGenes = Seq1(new UniqueGeneIdentifier(
+            GeneType.Fodder,
+            GeneIdentifier.New("gene:acme/acme-tools/1.0:test-fodder"),
+            Architecture.New(architecture)));
+
+        var result = CatletFeeding.Feed(config, resolvedGenes, _genepoolReaderMock.Object);
         
         var error = result.Should().BeLeft().Subject;
         error.Message.Should().Be("Could not expand the fodder gene 'gene:acme/acme-tools/1.0:test-fodder'.");
